refactor(hir-ty): extract member operations
